requirements-dev.txt
-r src/requirements.txt
pytest
flake8
coverage
pytest-cov
ruff
pre-commit
black
playwright
pytest-playwright
pytest-flask
axe-core-python
